functions/process-call-summary/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
